// src/app/models/IFD/portafolioIFDMoliendaLiquidar.ts
import { number } from "echarts";

export class PortafolioIFDMoliendaLiquidar{

  s212_Operacion: number;
  s212_FechaPacto: Date;
  s212_DescUnderlying: String;
  s212_Ticker: String;
  s212_FechaExpiracion: Date;
  s212_Instrumento: String;
  s212_TipoInstrumento: String;
  s212_CompraVenta: String;
  s212_Estado: String;
  s212_TipoLiquidacion: String;
  s212_CodigoBroker: String;
  s212_Sociedad: String;
  s212_NumeroContratos: number;
  s212_Bushels: number;
  s212_PrecioEjercicio: number;
  s212_Prima: number;
  s212_Broker: String;
  s212_MesContrato: String;
  s212_Costo: number;
  s212_CodigoEstrategia: number;
  s212_Estrategia: String;
  s212_MesEstrategia: String;
  s212_TipoCobertura: String;
  s212_M2M: number;
  s212_Ficha: String;
  s212_CuentaBroker: String;
  s212_Barrera: number;
  s212_TipoBarrera: String;
  s212_IdTipoOperacion: string;
  s212_TipoOperacion: String;
  s212_Base: String;
  s212_AcumulacionEstrategia: number;
  s212_elegir: Boolean;



}